+++ internal/aws/rum/app_monitor_resource_gen.go
@@ -672,7 +672,6 @@ func appMonitorResource(ctx context.Context) (resource.Resource, error) {
 		//	{
 		//	  "description": "The top-level internet domain names for which your application has administrative authority. The CreateAppMonitor requires either the domain or the domain list.",
 		//	  "items": {
-		//	    "description": "The top-level internet domain name for which your application has administrative authority. The CreateAppMonitor requires either the domain or the domain list.",
 		//	    "maxLength": 253,
 		//	    "minLength": 1,
 		//	    "pattern": "",